pyproject.toml
[project]
name = "doctor-system-with-ai-features"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
